refactor script tests
